edited ProdutionItem

+++ src/main/java/com/example/infraboxapi/productionItem/ProductionItem.java
@@ -23,7 +23,7 @@ public class ProductionItem {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String partName;
-    private Integer quantity;
+    private float quantity;
     @Column(name = "updated_on")
     private String updatedOn;
     @Column(name = "craeted_on")
